Refactor GaugeView invocation to prepare for multi-fragment layout
Depending on the screen size, selecting a Gauge from the Gauge list will
either start a new activity via an intent (for a small screen) or
alternatively just update existing fragments that are already displayed.

Fragments aren't supposed to explicitly know about each other (they should
be fluidly re-combineble) so the communication needs to go via the
'container' (ie the activity) which can make a decision about how to
dispatch the call.

There are few ways this can be done, for this approach I'm using this
tutorial as an example:

http://android-codelabs.appspot.com/resources/tutorials/honeypad/ex3.html