Auto-update manifest version name & code using android-maven-plugin
When publishing on the Android Market, and performing updates, you need to
increment the version code specified in the manifest:

http://developer.android.com/guide/publishing/versioning.html
http://developer.android.com/guide/publishing/publishing.html#marketupgrade

This change turns on manifest-update with the android-maven-plugin:

http://maven-android-plugin-m2site.googlecode.com/svn/manifest-update-mojo.html#manifest

The version name & code are auto-updated to reflect the pom version, which
as usual is incremented by the maven release process (has the nice
side-effect of git-tagging the code).

Note that the manifest files will be auto-formatted using JAXP transformer
defaults.