Revert "Remove unused location calculation"
Woah, I didn't realise getLocation() has side-effects - it modifies
the point passed to it, and returns the passed-in point. I guess
this is to avoid object allocation.

I did eyeball the airtraffic activity before removing that call,
but didn't realise that the rings were *just* appearing around location
of the last hit...!

This reverts commit ff54971f85fc8688e361810607c820e082d496e3.