Compute information needed to draw pins up front
Previously Hit objects were kept around for the life of the
pin shown on the map even though only the location and color 
were needed to draw it.  Now the Hit objects can be garbage
collected soon after they are processed and the x/y coordinates
and bitmaps can be persisted and redrawn when requested.