Launch gauge view activity from GaugeListFragment
This was previously done by propagating the click event from
the fragment to the activity so it would start the new activity
to view the gauge.

This removes the single use listener and still allows the activity
to intercept the intent request from the fragment by overriding
the startActivityFromFragment if needed.