Add automated signing-and-zipaligning for release
This is a required step for releasing an app to the Android Market - you
have to use a 'release' key rather than the debug key used for
development. Execute this command in the 'app' project:

$ mvn clean package -P release-to-market

This will produce an additional APK in the target dir with a
'-signed-aligned' suffix in the filename- that's the one you want to
release to market.

You'll need to have these system variables set:

github.sign.keystore - the location of your keystore file
github.sign.alias - the alias of the signing key, in our case 'github'
github.sign.storepass - the keystore password
github.sign.keypass - password used to unlock the 'github' key itself

http://developer.android.com/guide/publishing/app-signing.html
http://www.simpligility.com/2010/07/sign-zipalign-and-to-market-to-market-with-maven/
http://code.google.com/p/maven-android-plugin/wiki/SigningAPKWithMavenJarsigner