Reduce creation of NumberFormat objects
NumberFormat objects have to be treated in a thread-safe manner, one
nice thing about ViewHolders is that they are universally called in
a single-threaded fashion - they have to be, as the Android UI toolkit
isn't thread-safe and must always be manipulated on the UI thread. So
we're lucky and can get away with just having static NumberFormat
instances, reducing object creation when scrolling through lists.

http://developer.android.com/resources/articles/painless-threading.html