Stop list from jumping back to the top on reload
Uses setList() instead of setAdapter() - it's not a method available on
ListAdapter, so we have to ensure we have a ViewHoldingListAdapter.

This is a redo of commit 73689dc0 - which was broken because it left the
'Content' and 'Referrers' fragments blank. This was because a private
reference to the ListAdapter was being held in the ListLoadingFragment-
a bad idea because list fragments can actually have their ListAdapters
replaced when on coming back on screen, leaving the LLF holding an old
ref. Now, instead of holding our own ref to the ViewHoldingListAdapter,
we call getListAdapter() and cast it, ensuring we have the correct
ListAdapter.




fix the jump fix